Handle returns from callbacks correctly.
When a before save/create/update or before validations callback returns
false, the filter chain is halted and the given block will not be
called. This resulted in save operations happening although errors were
present.

Now the return value of all callbacks is checked and the save/update
methods will abort and return false if a filter halted the chain.