Added refinemnets and removed monkey patches
Up from Ruby 2.0 there is support for refinements. This
is a better way to redefine or add functinality to
existing (core) classes. The changes here are motivated
due to the error, that the patches for Date and Time
have breaking influences to other gems.

The folder core_ext still exists with the two patched
classes Date and Time. Because this change will require
Ruby >= 2.0, users with older projects and Ruby versions
have still the possibility to remove the refinements by
not 'using' them in module CouchPotato. This will be
described in the README in more detail.